pyproject.toml
[project]
name = "medi-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=23.2.1",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "fastapi>=0.111.0",
    "google-generativeai>=0.5.2",
    "httpx>=0.27.0",
    "numpy>=1.26.0",
    "openai>=1.30.0",
    "opencv-python-headless>=4.9.0",
    "pgvector>=0.2.5",
    "pillow>=10.3.0",
    "prometheus-client>=0.20.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
    "python-dotenv>=1.0.1",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "rapidfuzz>=3.9.0",
    "sqlalchemy[asyncio]>=2.0.29",
    "structlog>=24.1.0",
    "tenacity>=8.2.3",
    "uvicorn[standard]>=0.29.0",
]
